+++ app/models/concerns/whittaker_tech/midas/bankable.rb
@@ -133,7 +133,7 @@ module WhittakerTech
 
           define_method("#{name}_amount")   { public_send(name)&.amount }
           define_method("#{name}_format")   { public_send(name)&.amount&.format }
-          define_method("#{name}_in")       { |to| public_send(name)&.exchange_to(to)&.format }
+          # define_method("#{name}_in")       { |to| public_send(name)&.exchange_to(to)&.format }
 
           # Sets the coin value with the specified amount and currency.
           #